Updates the udev rules for local SSD disks. (#9)
The current rules only allow for symlinks of up to 8 local SSD disks,
and also rely on the NVMe controller for the local drives being labeled
as nvme0*.  This updates the rule to check for the appropriate
model name "nvme_card", which always corresponds to the local SSD
controller.

The additional PROGRAM element is required to 0-index the nsid so
that this change doesn't alter the expected symlinks (a value of 0 for
a namespace ID is invalid according to the spec).