Update udev rules and dependences for PD-NVMe disk ID support.
This commit updates the disk naming rules for PD-NVMe cards to generate symlinks if the underlying NVMe device supports it (via the vendor extension).  If the device lacks this support the rules will log an error and fail gracefully.

The included google_nvme_id is a bash script used to generate the ID_SERIAL and ID_SERIAL_SHORT environment variables used by udev.

The dependencies for debian and CentOS/RHEL were also updated to include nvme-cli, as google_nvme_id has a strict dependency on this package.

Finally, the disk-removal rules were updated to include NVMe in anticipation of the hot-detach feature for PD-NVMe.