Updated udev rules (#11)
* Update udev rules and dependences for PD-NVMe disk ID support.

This commit updates the disk naming rules for PD-NVMe cards to generate symlinks if the underlying NVMe device supports it (via the vendor extension).  If the device lacks this support the rules will log an error and fail gracefully.

The included google_nvme_id is a bash script used to generate the ID_SERIAL and ID_SERIAL_SHORT environment variables used by udev.

The dependencies for debian and CentOS/RHEL were also updated to include nvme-cli, as google_nvme_id has a strict dependency on this package.

Finally, the disk-removal rules were updated to include NVMe in anticipation of the hot-detach feature for PD-NVMe.

* Removed version restriction from spec file (el6 has its own spec file defined).

* Removes length limiting on xxd call (now reads to end instead of arbitrarily 256 characters).

Also fixes the following:
  - Various fixes with bash syntax (e.g., unnecessary {} around variable names)
  - Removed some dead code
  - Updated logic for clarity around verifying the device is a PD device

* Updates return value for error in function from 255 to 1

* Updates direct symlink gen to match udev rules

* Updated bash command for grabbing device name to make it more robust to whitespace

* Removed google_nvme_id script from centos 6 config since the nvme-cli package is not available

* Updated removal to fix typo and remove on buildroot side instead of repo