Fix NVMe partition names (#25)
* Fix the NVMe partition names

When partitions are created on NVMe devices, the current udev rules are using partition ID as the device ID to assign ID_SERIAL_SHORT. For example, if the first partition is created on NVMe devcie nvme0n5, the ID_SERIAL_SHORT of the partition will be local-nvme-ssd-0 intead of local-nvme-ssd-4. This PR will fix this issue.

* Update 65-gce-disk-naming.rules

Co-authored-by: Liam Hopkins <liamh@google.com>