Add the ability to signal the guest-agent's network events socket to google_set_hostname.sh
When this script is called without arguments (as it was used previously) it retains previous behavior and will still handle setting the hostname and FQDN.

This is part of handling hostname and FQDN setting in the guest-agent rather than the guest configs. The new architecture is that the user can configure the desired behavior regarding hostnames and FQDNs from the guest-agent, and the guest-configs simply signal to the guest-agent that some event on the system (such as an interface coming up) has happened.

The google_set_hostname script is used for triggering these events, the other scripts included here are hook scripts called by various networking stacks to send this signal.

* /etc/dhcp/dhclient.d/google_hostname.sh:
Distros: Enterprise Linux, Debian 10/11
Networking Stacks: dhclient, NetworkManager
The original architecture consisted of this dhclient exit hook calling the google_set_hostname script with arguments, now it calls the script to send the iface-up event. This hook is still used directly on debian 10 and 11 by dhclient. It is called indirectly by a compatibility script to call dhclient exit hooks from NetworkManager-dispatcher on enterprise linux.

* /etc/sysconfig/network/scripts/google_up.sh
Distros: SUSE
Network Stacks: wicked
This new script requires an associated config toggle in /etc/sysconfig/network/config to enable this script (POST_UP_SCRIPT="compat:suse:google_up.sh"). Unfortunately there is no way for packages to drop in configuration files either through /etc/sysconfig/network or /etc/wicked, so this line is appended in a post-install script.

* /usr/lib/networkd-dispatcher
Distros: Debian 12, Ubuntu
Network Stacks: systemd-networkd
This script is called when interfaces configured by systemd-networkd are routable. No additional configuration required, simply drop the script in the right folder.