gce-nic-naming: exit 1 so that udev ignores the rule on error
An issue with the prior behavior is that on error, we assume it is OK to use the predictable NIC name rather than allowing udev to decide what the fallback would be. When a guest configures "ethX" names and this script exits with errors, the ethX configuration is ignoring. This is particularly relevant in situations such as c3metal where we expect the script to error because we have an IDPF NIC but no accelerator devices.