Avoid duplicate entries for the metadata server in /etc/hosts
In cases where the metadata server /etc/hosts entry was not added by the
guest environment scripts there will be no "# Added by Google" marker. As
such the script will add a second reference to the metadata if it already
exists. We should only have 1 entry. Respect the already existing entry and
do not add a second entry.